Separate HTML utils from time utils
