+++ client/src/main/scala/walfie/gbf/raidfinder/client/Application.scala
@@ -6,6 +6,7 @@ import org.scalajs.dom
 import com.momentjs.Moment
 import scala.scalajs.js
 import walfie.gbf.raidfinder.protocol._
+import walfie.gbf.raidfinder.client.util.time._
 
 import dom.raw._
 import js.JSApp
@@ -13,38 +14,15 @@ import js.JSApp
 object Application extends JSApp {
   def main(): Unit = {
     val url = "ws://localhost:9000/ws/raids"
+    val bossTtl = Duration.hours(6)
+
     val websocket = new BinaryProtobufWebSocketClient(url)
     val client = new WebSocketRaidFinderClient(
-      websocket,
-      dom.window.localStorage,
-      Util.Duration.hours(6), // TODO: Don't hardcode
-      Util.SystemClock
+      websocket, dom.window.localStorage, bossTtl, SystemClock
     )
     client.updateBosses()
 
-    // TODO: Put this somewhere else
-    Moment.defineLocale(
-      "en-short",
-      js.Dictionary(
-        "parentLocale" -> "en",
-        "relativeTime" -> js.Dictionary(
-          "future" -> "in %s",
-          "past" -> "%s ago",
-          "s" -> "now",
-          "ss" -> "%ss",
-          "m" -> "1m",
-          "mm" -> "%dm",
-          "h" -> "1h",
-          "hh" -> "%dh",
-          "d" -> "1d",
-          "dd" -> "%dd",
-          "M" -> "1M",
-          "MM" -> "%dM",
-          "y" -> "1Y",
-          "yy" -> "%dY"
-        )
-      )
-    )
+    Moment.defineLocale("en-short", MomentShortLocale)
 
     val notification = new views.SnackbarNotification
