+++ client/src/main/scala/walfie/gbf/raidfinder/client/util/HtmlHelpers.scala
@@ -1,12 +1,11 @@
-package walfie.gbf.raidfinder.client
+package walfie.gbf.raidfinder.client.util
 
-import java.util.Date
 import org.scalajs.dom.raw._
 import org.scalajs.dom.{console, document}
 import scala.annotation.tailrec
 import scala.util.{Success, Failure, Try}
 
-object Util {
+object HtmlHelpers {
   @tailrec
   def findParent(element: Element, predicate: Element => Boolean): Option[Element] = {
     if (predicate(element)) Some(element)
@@ -49,18 +48,6 @@ object Util {
     s.background = "transparent"
 
     textArea
-  }
-
-  // TODO: Move this somewhere better
-  trait Clock { def now(): Date }
-  object SystemClock extends Clock { def now(): Date = new Date() }
-
-  case class Duration(milliseconds: Long) extends AnyVal
-  object Duration {
-    def seconds(s: Long): Duration = Duration(s * 1000)
-    def minutes(m: Long): Duration = Duration(m * 60 * 1000)
-    def hours(h: Long): Duration = Duration(h * 3600 * 1000)
-    def days(d: Long): Duration = Duration(d * 24 * 3600 * 1000)
   }
 }
